fix(Participants): remove the behavior when the binding target's converts the rating to an integer in the source property

+++ WorldFestSolution.XamarinApp/WorldFestSolution.XamarinApp/ViewModels/ParticipantsViewModel.cs
@@ -100,7 +100,7 @@ namespace WorldFestSolution.XamarinApp.ViewModels
             {
                 RaterId = Identity.Id,
                 UserId = (ratingBar.BindingContext as ParticipantUser).Id,
-                CountOfStars = ratingBar.SelectedStarValue,
+                CountOfStars = (int)ratingBar.SelectedStarValue,
                 IsRated = (ratingBar.BindingContext as ParticipantUser).IsRated
             };
             await UserRatingDataStore.AddItemAsync(rating);
